NoScriptPatch Compatibility Patch
The method of patching objects that don't have scripts in the vanilla game will actually over-write any script an other mod might have already added to it, causing several mods to not function properly.

To fix that, the .patch files for these items now use 2 batches.

The first batch checks if the path exists or not, and if not, creates the /scripts path. If the path already exist because an other mod added a script to it, that first batch will do nothing.

The second batch adds to this now-for-sure existing path the rl_crewanchorobject.lua script, either at the only script, or after an already patched-in script by an other mod.